Semester Schedule GET and POST endpoints created, plus additional refactoring

+++ src/main/java/com/davidwerfelmann/course_scheduler/data/InstructorRepository.java
@@ -3,5 +3,5 @@ package com.davidwerfelmann.course_scheduler.data;
 import com.davidwerfelmann.course_scheduler.models.Instructor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InstructorRepository extends JpaRepository<Instructor, Integer> {
+public interface InstructorRepository extends JpaRepository<Instructor, Long> {
 }
