+++ src/main/java/com/davidwerfelmann/course_scheduler/data/LocationRepository.java
@@ -3,5 +3,5 @@ package com.davidwerfelmann.course_scheduler.data;
 import com.davidwerfelmann.course_scheduler.models.Location;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocationRepository extends JpaRepository<Location, Integer> {
+public interface LocationRepository extends JpaRepository<Location, Long> {
 }
